Fix audio transcoding copy
